# apps/server/api/openapi/server-openapi.yaml
openapi: "3.0.3"
info:
  title: MCPist Server API
  version: "1.0.0"
  description: >-
    Internal REST API for the MCPist Go Server.
    Authenticated via X-Gateway-Token (Ed25519 JWT signed by Worker).
paths:
...
  /v1/me/register:
    post:
      operationId: registerUser
      summary: Register or find existing user from Clerk ID
      tags: [me]
      security:
        - gatewayToken: []
      responses:
        "200":
          description: User registered or already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResult"
        "400":
          description: Missing clerk_id or email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Profile ──────────────────────────────────────────────────
  /v1/me/profile:
    get:
      operationId: getMyProfile
      summary: Get current user profile
      tags: [me]
      security:
        - gatewayToken: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  securitySchemes:
    gatewayToken:
      type: apiKey
      in: header
      name: X-Gateway-Token
      description: Ed25519 JWT signed by Worker gateway
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterResult:
      type: object
      required: [id]
      properties:
        id:
          type: string

    # ── User ──
    UserProfile:
      type: object
      required: [id, account_status, plan_id, role, connected_count]
      properties:
        id:
          type: string
        account_status:
          type: string
        plan_id:
          type: string
        display_name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        role:
          type: string
        settings: {}
        connected_count:
          type: integer
...